refactor: extract fight state reducer and persistence
